Adding to plot script and dealing w/ initial supersaturation
Making some edits to the plotting script so it produces nicer plots.
Also noticed that the code doesn't properly deal with the situation where the initial vol_list/psurf/tsurf combination leads  to supersaturation at the surface. Made it so it just gets rid of the extra condensate and then recalculates psurf and vol_list with the formerly-supersaturated species now just saturated