Changed renormalization in case where vol_list != 1
I noticed the renormalization procedure when vol_list != 1 was wrong. It originally looped through all the volatiles and reset each to vol_list[vol]/sum(vol_list) in sequence; the problem with that is that as you changed vol_list[vol], the sum would change each step, so the next volatile would be divided by a different number. Fixed it so that it puts all the new volatiles fractions in a new dict and then transfers them back to the original dict after going through the whole list